ai-service/requirements.txt
cachetools==3.1.1
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
decorator==4.4.2
dictdiffer==0.9.0
distro==1.7.0
docutils==0.15.2
dpath==2.0.6
dulwich==0.20.37
dvc==1.11.16
elasticsearch==7.17.3
et-xmlfile==1.1.0
fastapi==0.87.0
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grandalf==0.6
graphviz==0.8.4
grpcio==1.27.2
h11==0.14.0
h5py==2.10.0
idna==2.8
imageio==2.8.0
importlib-metadata==1.5.0
imutils==0.5.4
iso8601==1.0.2
itsdangerous==1.1.0
jmespath==0.10.0
joblib==0.14.1
jsonpath-ng==1.5.3
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
mailchecker==4.1.16
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mxnet==1.9.1
numpy==1.18.1
opencv-python==4.2.0.32
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
Pillow==7.0.0
pycparser==2.19
pydantic==1.10.2
pydot==1.4.2
Pygments==2.12.0
pygtrie==2.3.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.15.7
PySocks==1.7.1
python-benedict==0.24.3
python-dateutil==2.8.1
python-fsutil==0.6.0
python-multipart==0.0.5
python-slugify==6.1.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.22.0
requests-oauthlib==1.3.0
rich==12.4.1
rsa==4.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
serial==0.0.97
shortuuid==1.0.9
shtab==1.5.4
sklearn==0.0
tensorboard==2.1.0
tensorflow==2.1.4
tensorflow-estimator==2.1.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.25.8
uvicorn==0.19.0

